requirements.txt
aiogram==3.23.0
aiohttp==3.13.2
APScheduler==3.11.1
python-dotenv==1.2.1
fluent.runtime==0.4.0
